[#60] Fix DOT by changing Health and MaxHealth to Float rather than Int.

+++ Assets/Scenes/scripts/Items/Item.cs
@@ -107,6 +107,6 @@ public class Item : MonoBehaviour
 
     // Events
     public virtual void OnPlayerAttack(Vector3 aimPosition, string targetTag){}
-    public virtual void OnPlayerHit(GameObject target, int amount){}
-    public virtual void OnPlayerGetHit(GameObject hitBy, int amount){}
+    public virtual void OnPlayerHit(GameObject target, float amount){}
+    public virtual void OnPlayerGetHit(GameObject hitBy, float amount){}
 }
